webapp/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.1
asttokens==3.0.0
attrs==24.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.57.0
imageio==2.37.0
ipykernel==6.29.5
ipynb-py-convert==0.4.6
ipython==8.18.0
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.14
kiwisolver==1.4.7
lazy_loader==0.4
langchain_community
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mistune==3.1.0
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
#pywin32==308
pyzmq==26.2.0
referencing==0.36.1
rpds-py==0.22.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.14

